When the card is swiped, the balance of the meal card is reduced.

+++ Education.BusinessLayer/Abstract/ICafeteriaCardService.cs
@@ -5,6 +5,6 @@ namespace Education.BusinessLayer.Abstract
 {
     public interface ICafeteriaCardService : IGenericService<CafeteriaCard>
     {
-        Task<CardTransactionResult> DeductBalanceAsync(int mealCardId, int amount);
+        Task<CardTransactionResult> DeductBalanceAsync(long mealCardId, int amount);
     }
 }
